Restructuring code
