+++ explainaboard/explainaboard_main.py
@@ -47,21 +47,21 @@ def main():
 
     # fine grained analysis
     if task == "absa":
-        explainaboard.tasks.absa.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.absa.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
     elif task == "ner":
-        explainaboard.tasks.ner.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.ner.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
     elif task == "pos":
-        explainaboard.tasks.pos.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.pos.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
     elif task == "chunk":
-        explainaboard.tasks.chunk.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.chunk.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
     elif task == "cws":
-        explainaboard.tasks.cws.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.cws.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
     elif task == "tc":
-        explainaboard.tasks.tc.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.tc.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
     elif task == "nli":
-        explainaboard.tasks.nli.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.nli.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
     elif task == "re":
-        explainaboard.tasks.re.tensoreval.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
+        explainaboard.tasks.re.eval_spec.evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
 
 
 
@@ -71,4 +71,4 @@ if __name__ == '__main__':
     # python explainaboard_main.py --task absa  --systems ./test-laptop.tsv --output ./output/a.json
     # python explainaboard_main.py --task ner --systems ./test-conll03.tsv --output ./a.json
     # python explainaboard_main.py --task re --systems ./test_re.tsv --output ./a.json
-    # python tensoreval.py  --task re --case True --systems ./test_re.tsv --output a.json --ci True
+    # python eval_spec.py  --task re --case True --systems ./test_re.tsv --output a.json --ci True
