+++ explainaboard/tasks/ner/__init__.py
@@ -1,2 +1,2 @@
 
-from . import tensoreval
+from . import eval_spec
