+++ explainaboard/tasks/tc/__init__.py
@@ -1,1 +1,1 @@
-from . import tensoreval
+from . import eval_spec
